Add method for cleaning up auth params
